+++ Tests/Public/User/Test-RunningAsAdmin.Tests.ps1
@@ -1,0 +1,5 @@
+﻿Describe 'Test-RunningAsAdmin' {
+    It 'It finds that current session is elevated' {
+        Test-RunningAsAdmin | Should -BeExactly $true
+    }
+}