+++ CHANGELOG.md
@@ -9,8 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 
-- Convert-ExcelSheetToJson
-  - Converts an Excel sheet from a workbook to JSON.
 - Get-OfficeUpdateChannel
   - Get the current Office Update Channel for Microsoft Office 365.
 - Set-OfficeUpdateChannel
@@ -19,6 +17,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
   - Add a module to RequiredModules in a PowerShell manifest. Add-RequiredModule
 uses Find-Module to find the module name, current version and GUID of a module
 and adds it to the RequiredModules section in the PowerShell manifest.
+- SwissKnife module dependencies
+  - Configuration
+  - ImportExcel
+  - PSScriptTools
+- Pester Tests
+  - Test-RunningAsAdmin
 
 ## [0.1.0] - 08/11-2019